move viewport config to addon
