+++ tests/unit/initializers/viewport-config-test.js
@@ -1,4 +1,5 @@
-import Ember from 'ember';
+import Application from '@ember/application';
+import { run } from '@ember/runloop';
 import ViewportConfigInitializer from 'dummy/initializers/viewport-config';
 import { module, test } from 'qunit';
 
@@ -8,9 +9,10 @@ let container, application;
 
 module('Unit | Initializer | viewport config', {
   beforeEach() {
-    Ember.run(function() {
-      application = Ember.Application.create();
+    run(function() {
+      application = Application.create();
       container = application.__container__;
+      application.register('config:environment', {});
       application.deferReadiness();
     });
   }
